+++ lib/src/extensions/message_delivery/CustomInterface.dart
@@ -1,0 +1,6 @@
+import 'package:xmpp_stone/src/elements/XmppElement.dart';
+
+abstract class CustomInterface {
+  CustomInterface addCustom(String customString);
+  XmppElement getCustom();
+}
