+++ lib/src/extensions/message_delivery/ReceiptInterface.dart
@@ -1,0 +1,9 @@
+import 'package:xmpp_stone/xmpp_stone.dart';
+
+enum ReceiptRequestType { NONE, REQUEST, RECEIVED }
+
+abstract class ReceiptInterface {
+  ReceiptInterface addRequestReceipt();
+  ReceiptInterface addReceivedReceipt();
+  XmppElement getReceipt();
+}
