+++ lib/src/extensions/message_delivery/TimeInterface.dart
@@ -1,0 +1,6 @@
+import 'package:xmpp_stone/src/elements/XmppElement.dart';
+
+abstract class TimeInterface {
+  TimeInterface addTime(int timeMilliseconds);
+  XmppElement getTime();
+}
