+++ lib/src/messages/MessageHandler.dart
@@ -5,6 +5,7 @@ import 'package:xmpp_stone/src/Connection.dart';
 import 'package:xmpp_stone/src/data/Jid.dart';
 import 'package:xmpp_stone/src/elements/stanzas/AbstractStanza.dart';
 import 'package:xmpp_stone/src/elements/stanzas/MessageStanza.dart';
+import 'package:xmpp_stone/src/extensions/message_delivery/ReceiptInterface.dart';
 import 'package:xmpp_stone/src/messages/MessageApi.dart';
 import 'package:xmpp_stone/xmpp_stone.dart';
